contrib-kubernetes: contrib-kubectl -> contrib-kubernetes
Because it isn't consistent with the rest of contribs.
kubernetes is the name of the tool and kubectl is the name of the cli
tool to interact with kubernetes.

Taking elixir in comparison, elixir is the langage name and mix is the build
tool. Most of the module consists of aliases over mix but the module is
named elixir which is good since the target is to add helpers to work
with elixir.