Improve Kubernetes integration for prezto (#5)
* contrib-kubernetes: contrib-kubectl -> contrib-kubernetes

Because it isn't consistent with the rest of contribs.
kubernetes is the name of the tool and kubectl is the name of the cli
tool to interact with kubernetes.

Taking elixir in comparison, elixir is the langage name and mix is the build
tool. Most of the module consists of aliases over mix but the module is
named elixir which is good since the target is to add helpers to work
with elixir.

* contrib-kubernetes: Add multiple helpers for Kubernetes

- Change every aliases to match kb prefix
- Add new helpful prefixes

* contrib-prompt: Add sorin-contrib prompt theme

This new theme is basically the sorin theme from prezto core with
kubernetes informations added in the prompt. I managed to write this
function file with minimum duplication with the original theme by
sourcing and reusing its functions. This did not reduced much the size
of the file though.