[kubernetes] Changed the way of customizing cluster formatting
In order to simplify the regular expression to be used to differenciate
prod clusters from dev clusters and remove the grep -P occurence, we
changed the configuration like this:
- if prod-clusters is defined, then clusters which have at least one
matching part with the regular expression will be colorized as red, the
others will be colorized as white.
- if prod-clusters is not defined and dev-clusters is defined, then
clusters which have at least one matching part with the regular
expression will be colorized as white, the others will be colorized as
red.
- if either prod-clusters and dev-clusters aren't defined then
dev-clusters-default will be used as dev-clusters.

This way, we removed the need of using perl regular expression and makes
the prompt compatible with macos and bsd like distros and users
now have two ways of defining their clusters.

Moreover, this feature will have minimal impact on existing users.